@@ -2,60 +2,87 @@
 Secure session management for Monarch Money MCP Server using keyring.
 """
 
-import keyring
 import logging
 import os
-from typing import Optional
-from monarchmoney import MonarchMoney, MonarchMoneyEndpoints
+from pathlib import Path
 
-# PATCH: Monarch Money rebranded from monarchmoney.com to monarch.com
-# The library hasn't been updated yet (as of v0.1.15), so we monkey-patch the BASE_URL
-# See: https://github.com/hammem/monarchmoney/issues/184
-MonarchMoneyEndpoints.BASE_URL = "https://api.monarch.com"
+from monarchmoney import MonarchMoney
+
+# Try to import keyring, but make it optional for container deployments
+try:
+    import keyring
+
+    KEYRING_AVAILABLE = True
+except ImportError:
+    KEYRING_AVAILABLE = False
 
 logger = logging.getLogger(__name__)
 
-# Keyring service identifiers
 # Keyring service identifiers
 KEYRING_SERVICE = "com.mcp.monarch-mcp-server"
 KEYRING_USERNAME = "monarch-token"
 
 # Standardize session file location to user's home directory
-from pathlib import Path
 DEFAULT_SESSION_FILE = Path.home() / ".mm" / "mm_session.pickle"
 
 
 class SecureMonarchSession:
     """Manages Monarch Money sessions securely using the system keyring."""
 
-    def save_token(self, token: str) -> None:
-        """Save the authentication token to the system keyring."""
+    def save_token(self, token: str) -> bool:
+        """Save the authentication token to the system keyring.
+
+        Returns:
+            True if token was saved successfully, False otherwise.
+        """
+        if not KEYRING_AVAILABLE:
+            logger.info("⏭️ Keyring not available, skipping token save to keyring")
+            return False
+
         try:
             keyring.set_password(KEYRING_SERVICE, KEYRING_USERNAME, token)
             logger.info("✅ Token saved securely to keyring")
 
             # Clean up any old insecure files
             self._cleanup_old_session_files()
+            return True
 
         except Exception as e:
-            logger.error(f"❌ Failed to save token to keyring: {e}")
-            raise
+            logger.error(
+                f"Failed to save token to keyring: {e}. "
+                "Ensure keyring backend is configured. "
+                "On Linux, install 'secretstorage' or configure gnome-keyring."
+            )
+            return False
 
-    def load_token(self) -> Optional[str]:
-        """Load the authentication token from querying keyring or session file."""
-        try:
-            # 1. Try keyring first
-            token = keyring.get_password(KEYRING_SERVICE, KEYRING_USERNAME)
-            if token:
-                logger.info("✅ Token loaded from keyring")
-                return token
-        except Exception as e:
-            logger.error(f"❌ Failed to load token from keyring: {e}")
+    def load_token(self) -> str | None:
+        """Load the authentication token from environment, keyring, or session file.
+
+        Priority order:
+        1. MONARCH_TOKEN environment variable (for cloud deployment)
+        2. System keyring (for local use)
+        3. Pickle file fallback (legacy support)
+        """
+        # 1. Try environment variable first (for cloud/container deployment)
+        env_token = os.getenv("MONARCH_TOKEN")
+        if env_token:
+            logger.info("✅ Token loaded from MONARCH_TOKEN environment variable")
+            return env_token
+
+        if KEYRING_AVAILABLE:
+            try:
+                # 2. Try keyring second (local use)
+                token = keyring.get_password(KEYRING_SERVICE, KEYRING_USERNAME)
+                if token:
+                    logger.info("✅ Token loaded from keyring")
+                    return token
+            except Exception as e:
+                logger.warning(f"⚠️ Failed to load token from keyring: {e}")
 
         # 2. Fallback to pickle file (standard library behavior)
         try:
             import pickle
-            
+
             if DEFAULT_SESSION_FILE.exists():
                 with open(DEFAULT_SESSION_FILE, "rb") as f:
                     data = pickle.load(f)
@@ -71,15 +98,16 @@ class SecureMonarchSession:
 
     def delete_token(self) -> None:
         """Delete the authentication token from system keyring and session file."""
-        try:
-            keyring.delete_password(KEYRING_SERVICE, KEYRING_USERNAME)
-            logger.info("🗑️ Token deleted from keyring")
-        except Exception as e:
-            error_type = type(e).__name__
-            if "PasswordDeleteError" in error_type or "not found" in str(e).lower():
-                logger.info("🔍 No token found in keyring to delete")
-            else:
-                logger.error(f"❌ Failed to delete token from keyring: {e}")
+        if KEYRING_AVAILABLE:
+            try:
+                keyring.delete_password(KEYRING_SERVICE, KEYRING_USERNAME)
+                logger.info("🗑️ Token deleted from keyring")
+            except Exception as e:
+                error_type = type(e).__name__
+                if "PasswordDeleteError" in error_type or "not found" in str(e).lower():
+                    logger.info("🔍 No token found in keyring to delete")
+                else:
+                    logger.warning(f"⚠️ Failed to delete token from keyring: {e}")
 
         # Also clean up session files
         self._cleanup_old_session_files()
@@ -92,7 +120,7 @@ class SecureMonarchSession:
         except Exception as e:
             logger.error(f"❌ Failed to delete mm_session.pickle: {e}")
 
-    def get_authenticated_client(self) -> Optional[MonarchMoney]:
+    def get_authenticated_client(self) -> MonarchMoney | None:
         """Get an authenticated MonarchMoney client."""
         # First try to load session using library's native persistence (loads cookies!)
         mm = MonarchMoney()
@@ -100,7 +128,9 @@ class SecureMonarchSession:
             # Explicitly use the default session file path
             mm.load_session(filename=str(DEFAULT_SESSION_FILE))
             if mm.token:
-                logger.info("✅ MonarchMoney client loaded with native session (cookies+token)")
+                logger.info(
+                    "✅ MonarchMoney client loaded with native session (cookies+token)"
+                )
                 return mm
         except Exception as e:
             logger.warning(f"⚠️ Failed to load native session: {e}")
@@ -110,11 +140,13 @@ class SecureMonarchSession:
         if token:
             try:
                 client = MonarchMoney(token=token)
-                logger.info("✅ MonarchMoney client created with stored token (token-only)")
+                logger.info(
+                    "✅ MonarchMoney client created with stored token (token-only)"
+                )
                 return client
             except Exception as e:
                 logger.error(f"❌ Failed to create MonarchMoney client: {e}")
-        
+
         return None
 
     def save_authenticated_session(self, mm: MonarchMoney) -> None:
@@ -122,13 +154,15 @@ class SecureMonarchSession:
         if mm.token:
             # 1. Save to keyring (most secure)
             self.save_token(mm.token)
-            
+
             # 2. Save native session pickle (persists cookies for long sessions)
             try:
                 # Be sure to create the directory if it doesn't exist
                 DEFAULT_SESSION_FILE.parent.mkdir(parents=True, exist_ok=True)
                 mm.save_session(filename=str(DEFAULT_SESSION_FILE))
-                logger.info(f"✅ Full session (cookies) saved to {DEFAULT_SESSION_FILE}")
+                logger.info(
+                    f"✅ Full session (cookies) saved to {DEFAULT_SESSION_FILE}"
+                )
             except Exception as e:
                 logger.warning(f"⚠️ Could not save native session pickle: {e}")
         else:
@@ -136,11 +170,11 @@ class SecureMonarchSession:
 
     def _cleanup_old_session_files(self) -> None:
         """Clean up old insecure session files.
-        
+
         Uses absolute paths based on user's home directory to avoid
         accidentally cleaning up files in wrong working directories.
         """
-        
+
         cleanup_paths = [
             # home / ".mm" / "mm_session.pickle",  <-- KEEP THIS ONE! (Handled by delete_token now)
             Path.home() / "monarch_session.json",
