@@ -18,7 +18,9 @@ class MonarchMCPError(Exception):
 class AuthenticationError(MonarchMCPError):
     """Raised when authentication fails or session is invalid."""
 
-    def __init__(self, message: str = "Authentication required", details: str | None = None):
+    def __init__(
+        self, message: str = "Authentication required", details: str | None = None
+    ):
         super().__init__(message, details)
 
 
@@ -26,7 +28,9 @@ class SessionExpiredError(AuthenticationError):
     """Raised when the session has expired and needs refresh."""
 
     def __init__(self, details: str | None = None):
-        super().__init__("Session expired. Please run login_setup.py to re-authenticate.", details)
+        super().__init__(
+            "Session expired. Please run login_setup.py to re-authenticate.", details
+        )
 
 
 class NetworkError(MonarchMCPError):
@@ -39,7 +43,9 @@ class NetworkError(MonarchMCPError):
 class APIError(MonarchMCPError):
     """Raised when the Monarch Money API returns an error."""
 
-    def __init__(self, message: str, status_code: int | None = None, details: str | None = None):
+    def __init__(
+        self, message: str, status_code: int | None = None, details: str | None = None
+    ):
         self.status_code = status_code
         super().__init__(message, details)
 
@@ -47,7 +53,9 @@ class APIError(MonarchMCPError):
 class ValidationError(MonarchMCPError):
     """Raised when input validation fails."""
 
-    def __init__(self, message: str, field: str | None = None, details: str | None = None):
+    def __init__(
+        self, message: str, field: str | None = None, details: str | None = None
+    ):
         self.field = field
         super().__init__(message, details)
 
@@ -55,7 +63,11 @@ class ValidationError(MonarchMCPError):
 class SafetyError(MonarchMCPError):
     """Raised when a safety check blocks an operation."""
 
-    def __init__(self, message: str = "Operation blocked by safety check", details: str | None = None):
+    def __init__(
+        self,
+        message: str = "Operation blocked by safety check",
+        details: str | None = None,
+    ):
         super().__init__(message, details)
 
 
@@ -65,14 +77,5 @@ class EmergencyStopError(SafetyError):
     def __init__(self):
         super().__init__(
             "🚨 EMERGENCY STOP ACTIVE: All write operations disabled.",
-            "Use disable_emergency_stop to re-enable."
+            "Use disable_emergency_stop to re-enable.",
         )
-
-
-class RateLimitError(MonarchMCPError):
-    """Raised when rate limits are exceeded."""
-
-    def __init__(self, operation: str, limit: str, details: str | None = None):
-        self.operation = operation
-        self.limit = limit
-        super().__init__(f"Rate limit exceeded for {operation}: {limit}", details)
